refactor(handler): split admin, agents files under 300 line limit
